requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
pip==21.3.1
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
setuptools==58.1.0
urllib3==1.26.7
